refactor(15311): relocate component
